[REF] mail: move channel commands to discuss folder
Part of task-3265211
