+++ addons/mail/static/src/core/thread_service.js
@@ -879,13 +879,6 @@ export class ThreadService {
      * @param {string} body
      */
     async post(thread, body, { attachments = [], isNote = false, parentId, rawMentions }) {
-        const command = this.store.user
-            ? this.messageService.getCommandFromText(thread, body)
-            : undefined;
-        if (command) {
-            await this.executeCommand(thread, command, body);
-            return;
-        }
         let tmpMsg;
         const subtype = isNote ? "mail.mt_note" : "mail.mt_comment";
         const validMentions = this.store.user
