+++ addons/mail/static/src/discuss/channel_commands.js
