+++ addons/mail/static/tests/suggestion/suggestion_tests.js
@@ -129,54 +129,6 @@ QUnit.test("select @ mention insert mention text in composer", async (assert) =>
     assert.strictEqual($(".o-mail-Composer-input").val().trim(), "@TestPartner");
 });
 
-QUnit.test('display command suggestions on typing "/"', async (assert) => {
-    const pyEnv = await startServer();
-    const channelId = pyEnv["discuss.channel"].create({
-        name: "General",
-        channel_type: "channel",
-    });
-    const { openDiscuss } = await start();
-    await openDiscuss(channelId);
-    assert.containsNone($, ".o-mail-Composer-suggestionList .o-open");
-    await insertText(".o-mail-Composer-input", "/");
-    assert.containsOnce($, ".o-mail-Composer-suggestionList .o-open");
-});
-
-QUnit.test("use a command for a specific channel type", async (assert) => {
-    const pyEnv = await startServer();
-    const channelId = pyEnv["discuss.channel"].create({ channel_type: "chat" });
-    const { openDiscuss } = await start();
-    await openDiscuss(channelId);
-    assert.containsNone($, ".o-mail-Composer-suggestionList .o-open");
-    assert.strictEqual($(".o-mail-Composer-input").val(), "");
-    await insertText(".o-mail-Composer-input", "/");
-    await click(".o-mail-Composer-suggestion");
-    assert.strictEqual(
-        $(".o-mail-Composer-input").val().replace(/\s/, " "),
-        "/who ",
-        "command + additional whitespace afterwards"
-    );
-});
-
-QUnit.test(
-    "command suggestion should only open if command is the first character",
-    async (assert) => {
-        const pyEnv = await startServer();
-        const channelId = pyEnv["discuss.channel"].create({
-            name: "General",
-            channel_type: "channel",
-        });
-        const { openDiscuss } = await start();
-        await openDiscuss(channelId);
-        assert.containsNone($, ".o-mail-Composer-suggestionList .o-open");
-        assert.strictEqual($(".o-mail-Composer-input").val(), "");
-        await insertText(".o-mail-Composer-input", "bluhbluh ");
-        assert.strictEqual($(".o-mail-Composer-input").val(), "bluhbluh ");
-        await insertText(".o-mail-Composer-input", "/");
-        assert.containsNone($, ".o-mail-Composer-suggestionList .o-open");
-    }
-);
-
 QUnit.test('display channel mention suggestions on typing "#"', async (assert) => {
     const pyEnv = await startServer();
     const channelId = pyEnv["discuss.channel"].create({
